+++ kensu/utils/helpers.py
@@ -1,7 +1,7 @@
 import re
 from hashlib import sha1
 
-
+from kensu.utils.kensu_provider import KensuProvider
 
 
 def to_snake_case(name):
@@ -25,7 +25,6 @@ def to_hash_key(o):
 
 
 def eventually_report_in_mem(o):
-    from kensu.utils.kensu_provider import KensuProvider
     kensu = KensuProvider().instance()
     if kensu.report_in_mem:
         o._report()
@@ -43,7 +42,6 @@ def get_absolute_path(path):
 
 
 def maybe_report(o, report):
-    from kensu.utils.kensu_provider import KensuProvider
     dam = KensuProvider().instance()
     if report:
         o._report()
@@ -58,3 +56,24 @@ def extract_ksu_ds_schema(kensu, orig_variable, report=False, register_orig_data
     if register_orig_data:
         kensu.real_schema_df[schema.to_guid()] = orig_variable
     return ds, schema
+
+
+def report_all2all_lineage(in_obj, out_obj, op_type, in_inmem=True, out_inmem=True):
+    kensu = KensuProvider().instance()
+    in_ds, in_schema = extract_ksu_ds_schema(kensu,
+                                             in_obj,
+                                             report=kensu.report_in_mem or not in_inmem,
+                                             register_orig_data=not in_inmem)
+    out_ds, out_schema = extract_ksu_ds_schema(kensu,
+                                               out_obj,
+                                               report=kensu.report_in_mem or not out_inmem,
+                                               register_orig_data=not out_inmem)
+    logging.debug("in_schema=" + str(in_schema) + "\nout_schema=" + str(out_schema))
+    if kensu.mapping:
+        for col in out_obj:
+            if col in [v.name for v in in_schema.pk.fields]:
+                kensu.add_dependencies_mapping(guid=out_schema.to_guid(),
+                                               col=str(col),
+                                               from_guid=in_schema.to_guid(),
+                                               from_col=str(col),
+                                               type=op_type)
